# Improve performance of assert_key_data