@mitchelloharawild 2019-08-14
Any reason why the row names of a tsibble is required? Removing it will speed up `new_tsibble` by ~40%.

https://github.com/tidyverts/tsibble/blob/2798909f19d4fc6534b641c331bf2cc4ed99b458/R/as-tsibble.R#L353