@earowang 2019-08-14
I suppose it can be safely removed for now. `new_tibble()` has that line.